update tx processing to use raw sql
Entity framework seems to not be able to generate sql string to increment, or decrement a value, instead it will always just set the value to latest saved value.
This becomes a problem when many transactions are happening at the same time, they will be overwriting each others changes.
one solution to this is to generate sql string manually which increments / decrements the current value stored in the database.